+++ evcxr/src/lib.rs
@@ -29,7 +29,6 @@ mod crate_config;
 mod eval_context;
 #[allow(dead_code)]
 mod evcxr_internal_runtime;
-mod idents;
 mod item;
 mod module;
 mod runtime;
@@ -39,3 +38,14 @@ pub use crate::command_context::CommandContext;
 pub use crate::errors::{CompilationError, Error};
 pub use crate::eval_context::{EvalContext, EvalContextOutputs, EvalOutputs};
 pub use crate::runtime::runtime_hook;
+
+/// Return the directory that evcxr tools should use for their configuration.
+///
+/// By default this is the `evcxr` subdirectory of whatever `dirs::config_dir()`
+/// returns, but it can be overridden by the `EVCXR_CONFIG_DIR` environment
+/// variable.
+pub fn config_dir() -> Option<std::path::PathBuf> {
+    std::env::var_os("EVCXR_CONFIG_DIR")
+        .map(std::path::PathBuf::from)
+        .or_else(|| dirs::config_dir().map(|d| d.join("evcxr")))
+}