+++ RELEASE_NOTES.md
@@ -1,3 +1,22 @@
+# Version 0.4.7
+* Fixed segfault if executing code without variables followed by code with
+  variables.
+* Automatically use lld if it's installed. Use :linker command to override.
+* Handle variables that contain type inference placeholders `_`.
+* Added support for overriding the config dir with the environment variable
+  EVCXR_CONFIG_DIR (thanks Thom Chiovoloni).
+* Run prelude.rs after init.evcxr if one exists (thanks Thom Chiovoloni).
+* Allow output format to be specified. e.g. :fmt {:#?} (thanks κeen)
+
+# Version 0.4.6
+* Fixes to work with rust 1.41
+* Improved handling of cargo workspaces.
+* Now works if the user has overridden Cargo's default target directory (thanks
+  Aloxaf).
+* Fixed prompt color (thanks Dmitry Murzin).
+* Added a flag to disable use of readline (thanks Dmitry Murzin).
+* Fix for binder (thanks Luiz Irber)
+
 # Version 0.4.5
 * Escape reserved words "async" and "try" when encountered in types.
 * Use vendored ZMQ library by default.